+++ src/components/page/overview/FilterBar.tsx
@@ -1,5 +1,7 @@
 import { TableFilter } from '@/components/page/overview/TableFilter'
 import FilterFormModal from '@/components/page/overview/FilterFormModal'
+import { TableSort } from '@/components/page/overview/TableSort'
+import TableSortModal from '@/components/page/overview/TableSortModal'
 
 type FilterBarProps = {
   filters: TableFilter[]
@@ -7,6 +9,8 @@ type FilterBarProps = {
   updateFilter(filter: TableFilter): void
   currentEntries: number
   totalEntries: number
+  tableSort: TableSort | null
+  setTableSort(tableSort: TableSort | null): void
 }
 
 const FilterBar = ({
@@ -14,10 +18,14 @@ const FilterBar = ({
   removeFilter,
   updateFilter,
   currentEntries,
-  totalEntries
+  totalEntries,
+  tableSort,
+  setTableSort
 }: FilterBarProps) => (
   <div className="flex flex-row justify-between border-b border-gray-800 px-6 py-2.5">
     <div className="flex flex-row gap-3">
+      {tableSort && <TableSortModal tableSort={tableSort} setTableSort={setTableSort} />}
+      {tableSort && <div className="mx-3 h-full border-l border-gray-800" />}
       {filters.map((filter: TableFilter) => (
         <FilterFormModal filter={filter} removeFilter={removeFilter} updateFilter={updateFilter} />
       ))}
