+++ src/components/page/overview/TableSortModal.tsx
@@ -9,37 +9,37 @@ import {
 } from 'react-icons/bs'
 
 type TableSortModalProps = {
-  sort: TableSort
-  setSort: (sort: TableSort | null) => void
+  tableSort: TableSort
+  setTableSort: (sort: TableSort | null) => void
 }
 
-const TableSortModal = ({ sort, setSort }: TableSortModalProps) => {
+const TableSortModal = ({ tableSort, setTableSort }: TableSortModalProps) => {
   const textGray100 = 'text-gray-100'
   const textGray500 = 'text-gray-500'
   return (
     <div className="flex flex-col space-y-3 p-2">
       <div className="flex flex-row space-x-1">
-        <Button onClick={() => setSort(null)} variant="onlyIconNoBorderNoBG" Icon={RxCross2} />
-        <p className="text-14 text-gray-100">{sort.column}</p>
+        <Button onClick={() => setTableSort(null)} variant="onlyIconNoBorderNoBG" Icon={RxCross2} />
+        <p className="text-14 text-gray-100">{tableSort.column}</p>
       </div>
       <div className="flex flex-row space-x-1">
         <Button
-          onClick={() => setSort({ ...sort, direction: 'asc' })}
-          variant={sort.direction === 'asc' ? 'highlighted' : 'normal'}
+          onClick={() => setTableSort({ ...tableSort, direction: 'asc' })}
+          variant={tableSort.direction === 'asc' ? 'highlighted' : 'normal'}
           text="Asc."
-          Icon={sort.column === 'Name' ? BsSortAlphaUpAlt : BsSortNumericUp}
+          Icon={tableSort.column === 'Name' ? BsSortAlphaUpAlt : BsSortNumericUp}
           order="ltr"
-          iconColor={sort.direction === 'asc' ? textGray100 : textGray500}
-          textColor={sort.direction === 'asc' ? textGray100 : textGray500}
+          iconColor={tableSort.direction === 'asc' ? textGray100 : textGray500}
+          textColor={tableSort.direction === 'asc' ? textGray100 : textGray500}
         />
         <Button
-          onClick={() => setSort({ ...sort, direction: 'desc' })}
-          variant={sort.direction === 'desc' ? 'highlighted' : 'normal'}
+          onClick={() => setTableSort({ ...tableSort, direction: 'desc' })}
+          variant={tableSort.direction === 'desc' ? 'highlighted' : 'normal'}
           text="Desc."
-          Icon={sort.column === 'Name' ? BsSortAlphaDown : BsSortNumericDownAlt}
+          Icon={tableSort.column === 'Name' ? BsSortAlphaDown : BsSortNumericDownAlt}
           order="ltr"
-          iconColor={sort.direction === 'desc' ? textGray100 : textGray500}
-          textColor={sort.direction === 'desc' ? textGray100 : textGray500}
+          iconColor={tableSort.direction === 'desc' ? textGray100 : textGray500}
+          textColor={tableSort.direction === 'desc' ? textGray100 : textGray500}
         />
       </div>
     </div>
