+++ src/components/side-effects/Compare/index.tsx
@@ -90,6 +90,8 @@ const Compare = () => {
           updateFilter={updateFilter}
           currentEntries={filteredRowCount}
           totalEntries={data.length}
+          tableSort={tableSort}
+          setTableSort={setTableSort}
         />
       )}
